Upgraded Hasher to 64 bit. Be GONE 32 bit! Too much stress... never again... Also added +delete lable.

+++ hash_util/hashutil.cpp
@@ -25,7 +25,7 @@ support@lumpology.com
 using namespace std;
 
 //Someday I hope to give who ever wrote this public domain code the credit they desurve. However so far my search has proved to find no results.
-constexpr uint32_t UIRN(char const *s, unsigned int count) {
+constexpr uint64_t UIRN(char const *s, unsigned int count) {
 	return count ? (UIRN(s, count - 1) ^ s[count - 1]) * 16777619u : 2166136261u;
 }
 #define COREFC(X) UIRN(X, strlen(X))
